refactor(ui): move AppStateEntity to dbflux_ui_base, shim at old path (#102)
Defines AppStateEntity + AppStateChanged + AuthProfileCreated +
McpRuntimeEventRaised (mcp-gated) in dbflux_ui_base::app_state_entity.
Turns crates/dbflux_ui/src/app_state_entity.rs into a re-export shim so
the ~40 call-sites in dbflux_ui remain unchanged.

Fixes the pre-existing non-mcp build break in app.rs: McpRuntimeEventRaised
re-export is now guarded with #[cfg(feature = "mcp")].
