+++ crates/dbflux_ui_base/src/lib.rs
@@ -3,3 +3,9 @@
 //! This crate holds types that require both `dbflux_app` domain knowledge and
 //! GPUI integration, sitting between the pure `dbflux_components` layer and the
 //! full `dbflux_ui` crate.
+
+pub mod app_state_entity;
+
+#[cfg(feature = "mcp")]
+pub use app_state_entity::McpRuntimeEventRaised;
+pub use app_state_entity::{AppStateChanged, AppStateEntity, AuthProfileCreated};
